backend/requirements.txt
flask==2.3.3
flask_cors
flask_sockets==0.2.1
numpy==1.24.3
scipy==1.10.1
pydub==0.25.1
spotipy==2.23.0
yt-dlp==2023.7.6
python-dotenv==1.0.0
requests
websocket-client
flask-sock
gunicorn 
sounddevice